python-Django/requirements.txt
Django==4.2.4
mysqlclient==1.4.6
djangorestframework==3.14.0